Move vendor/ from src/ to package root + fix devtools::install()
Three fixes for the scaffolding/build system:

1. Move vendor directory from src/vendor/ to package root vendor/.
   pkgbuild::sources() recursively scans src/ for .rs files, so having
   thousands of vendored .rs files there triggers false rebuilds on every
   devtools::load_all(). Same reasoning as CARGO_TARGET_DIR being outside src/.

2. Configure now unpacks vendor.tar.xz in ALL modes (not just CRAN mode).
   This fixes devtools::install() which runs R CMD build (excluding vendor/
   via .Rbuildignore) then R CMD INSTALL from the tarball — configure needs
   to unpack the tarball regardless of NOT_CRAN setting.

3. Starter lib.rs templates use f64 instead of i32. R's bare numeric
   literals (3, 4) are REALSXP (double), so add(3, 4) with i32 params
   gives a confusing type error for new users.

+++ minirextendr/inst/templates/rpkg/lib.rs
@@ -14,14 +14,17 @@ use miniextendr_api::{miniextendr, miniextendr_module};
 //    The first install compiles your Rust code; devtools::document() then
 //    runs the compiled macros to emit R/miniextendr_wrappers.R; the second
 //    install bundles those new wrappers into the package.
+//
+// NOTE: `cargo check` won't work until ./configure has been run at least once,
+// because document.rs contains template placeholders that need to be resolved.
 
-/// A simple function that adds two integers
+/// A simple function that adds two numbers
 ///
-/// @param a First integer
-/// @param b Second integer
+/// @param a First number
+/// @param b Second number
 /// @return Sum of a and b
 #[miniextendr]
-pub fn add(a: i32, b: i32) -> i32 {
+pub fn add(a: f64, b: f64) -> f64 {
     a + b
 }
 
@@ -33,6 +36,34 @@ pub fn add(a: i32, b: i32) -> i32 {
 pub fn hello(name: &str) -> String {
     format!("Hello, {}!", name)
 }
+
+// ---- Classes ----
+//
+// You can expose Rust structs as R6 classes. Here's a simple example:
+//
+//   use miniextendr_api::ExternalPtr;
+//
+//   #[derive(ExternalPtr)]
+//   pub struct Counter {
+//       value: i32,
+//   }
+//
+//   #[miniextendr]
+//   impl Counter {
+//       pub fn new() -> Self {
+//           Counter { value: 0 }
+//       }
+//
+//       pub fn increment(&mut self) {
+//           self.value += 1;
+//       }
+//
+//       pub fn get(&self) -> i32 {
+//           self.value
+//       }
+//   }
+//
+// Then register with: `impl Counter;` in the miniextendr_module! below.
 
 // Register the module with R
 miniextendr_module! {
